MNIST-MLP working
