+++ examples/MNIST-Cryptonets/README.md
@@ -18,15 +18,15 @@ This trains the network briefly and stores the network weights.
 To test the network, with
   * encrypted data,
 ```
-NGRAPH_ENCRYPT_DATA=1 NGRAPH_BATCH_DATA=1 NGRAPH_HE_SEAL_CONFIG=../../test/model/he_seal_ckks_config_N13_L7.json NGRAPH_TF_BACKEND=HE_SEAL_CKKS python test.py --batch_size=4096
+NGRAPH_ENCRYPT_DATA=1 NGRAPH_BATCH_DATA=1 NGRAPH_HE_SEAL_CONFIG=../../test/model/he_seal_ckks_config_N13_L7.json NGRAPH_TF_BACKEND=HE_SEAL python test.py --batch_size=4096
 ```
  * encrypted model,
  ```
-NGRAPH_ENCRYPT_MODEL=1 NGRAPH_BATCH_DATA=1 NGRAPH_HE_SEAL_CONFIG=../../test/model/he_seal_ckks_config_N13_L7.json NGRAPH_TF_BACKEND=HE_SEAL_CKKS python test.py --batch_size=4096
+NGRAPH_ENCRYPT_MODEL=1 NGRAPH_BATCH_DATA=1 NGRAPH_HE_SEAL_CONFIG=../../test/model/he_seal_ckks_config_N13_L7.json NGRAPH_TF_BACKEND=HE_SEAL python test.py --batch_size=4096
 ```
 * encrypted data and model,
 ```
-NGRAPH_ENCRYPT_DATA=1 NGRAPH_ENCRYPT_MODEL=1 NGRAPH_BATCH_DATA=1 NGRAPH_HE_SEAL_CONFIG=../../test/model/he_seal_ckks_config_N13_L7.json NGRAPH_TF_BACKEND=HE_SEAL_CKKS python test.py --batch_size=4096
+NGRAPH_ENCRYPT_DATA=1 NGRAPH_ENCRYPT_MODEL=1 NGRAPH_BATCH_DATA=1 NGRAPH_HE_SEAL_CONFIG=../../test/model/he_seal_ckks_config_N13_L7.json NGRAPH_TF_BACKEND=HE_SEAL python test.py --batch_size=4096
 ```
 
 This runs inference on the Cryptonets network using the SEAL CKKS backend.
