+++ python/CMakeLists.txt
@@ -56,7 +56,7 @@ message("EXTERNAL_INSTALL_LIB_DIR ${EXTERNAL_INSTALL_LIB_DIR}")
 
 add_library(he_seal_client SHARED ${CMAKE_SOURCE_DIR}/src/seal/he_seal_client.cpp)
 # TODO: remove he_seal_backend dependencty?
-add_dependencies(he_seal_client ext_seal he_seal_backend he_seal_backend)
+add_dependencies(he_seal_client ext_seal he_seal_backend)
 target_link_libraries(he_seal_client seal libjson libboost)
 install(TARGETS he_seal_client DESTINATION ${HE_INSTALL_LIB_DIR})
