+++ src/seal/he_seal_backend.cpp
@@ -17,6 +17,7 @@
 #include <limits>
 #include <memory>
 
+#include "client_util.hpp"
 #include "he_plain_tensor.hpp"
 #include "he_seal_cipher_tensor.hpp"
 #include "ngraph/runtime/backend_manager.hpp"
